backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
mysqlclient==2.2.0
Pillow==10.0.1
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
celery==5.3.1
redis==5.0.0
pdfplumber==0.10.2
pytesseract==0.3.10
reportlab==4.0.4
gunicorn==21.2.0
python-decouple==3.8
 